Lottery/requirements/base.txt
arrow==0.13.1
asgiref==3.3.4
binaryornot==0.4.4
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
cryptography==3.4.7
Django==3.2.4
djangorestframework==3.12.4
idna==2.10
importlib-metadata==3.10.0
Jinja2==2.11.3
jinja2-time==0.2.0
MarkupSafe==1.1.1
pip==21.1.2
poyo==0.5.0
pycparser==2.20
PyMySQL==1.0.2
pyOpenSSL==20.0.1
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
uvicorn==0.14.0
celery==5.1.2
django-celery-beat==2.2.1
django-filter==2.4.0
redis==3.5.3
watchgod==0.7
django-celery-results==2.2.0
flower==1.0.0
gunicorn==20.1.0
